requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
Pillow==9.1.1
PyMuPDF==1.19.6
PyPDF2==1.28.4
pytz==2022.1
sqlparse==0.4.2
tk==0.1.0
tensorflow-cpu>=1.15.2
keras==2.3.1
imutils==0.5.3
numpy==1.18.2
opencv-python==4.2.0.*
matplotlib==3.2.1
scipy==1.4.1
